Refactor middlewares into their own directory.
Reduces the bulkiness a bit, especially since we're adding more.
