+++ src/client.ts
@@ -10,12 +10,13 @@ import {
   cjStateFactory,
   htmlStateFactory
 } from './state';
-import { parseContentType, isSafeMethod } from './http/util';
+import { parseContentType } from './http/util';
 import { resolve } from './util/uri';
 import { LinkVariables } from './link';
 import { FollowPromiseOne } from './follow-promise';
 import { StateCache, ForeverCache } from './cache';
-import * as LinkHeader from 'http-link-header';
+import cacheExpireMiddleware from './middlewares/cache-expire';
+import acceptMiddleware from './middlewares/accept-header';
 
 export default class Client {
 
@@ -67,8 +68,8 @@ export default class Client {
   constructor(bookmarkUri: string) {
     this.bookmarkUri = bookmarkUri;
     this.fetcher = new Fetcher();
-    this.fetcher.use( this.cacheExpireHandler );
-    this.fetcher.use( this.acceptHeader );
+    this.fetcher.use(cacheExpireMiddleware(this));
+    this.fetcher.use(acceptMiddleware(this));
     this.cache = new ForeverCache();
     this.resources = new Map();
   }
@@ -192,90 +193,5 @@ export default class Client {
     }
 
   }
-
-  private acceptHeader: FetchMiddleware = (request, next) => {
-
-    if (!request.headers.has('Accept')) {
-      const acceptHeader = Object.entries(this.contentTypeMap).map(
-        ([contentType, [stateFactory, q]]) => contentType + ';q=' + q
-      ).join(', ');
-      request.headers.set('Accept', acceptHeader);
-    }
-    return next(request);
-
-  };
-
-  private cacheExpireHandler: FetchMiddleware = async(request, next) => {
-
-    /**
-     * Prevent a 'stale' event from being emitted, but only for the main
-     * uri
-     */
-    let noStaleEvent = false;
-
-    if (request.headers.has('X-KETTING-NO-STALE')) {
-      noStaleEvent = true;
-      request.headers.delete('X-KETTING-NO-STALE');
-    }
-
-    const response = await next(request);
-    if (isSafeMethod(request.method)) {
-      return response;
-    }
-
-    if (!response.ok) {
-      // There was an error, no cache changes
-      return response;
-    }
-
-    // We just processed an unsafe method, lets notify all subsystems.
-    const expireUris = [];
-    if (!noStaleEvent && request.method !== 'DELETE') {
-      // Sorry for the double negative
-      expireUris.push(request.url);
-    }
-
-    // If the response had a Link: rel=invalidate header, we want to
-    // expire those too.
-    if (response.headers.has('Link')) {
-      for (const httpLink of LinkHeader.parse(response.headers.get('Link')!).rel('invalidates')) {
-        const uri = resolve(request.url, httpLink.uri);
-        expireUris.push(uri);
-      }
-    }
-
-    // Location headers should also expire
-    if (response.headers.has('Location')) {
-      expireUris.push(
-        resolve(request.url, response.headers.get('Location')!)
-      );
-    }
-    // Content-Location headers should also expire
-    if (response.headers.has('Content-Location')) {
-      expireUris.push(
-        resolve(request.url, response.headers.get('Content-Location')!)
-      );
-    }
-
-    for (const uri of expireUris) {
-      this.cache.delete(request.url);
-
-      const resource = this.resources.get(uri);
-      if (resource) {
-        // We have a resource for this object, notify it as well.
-        resource.emit('stale');
-      }
-    }
-    if (request.method === 'DELETE') {
-      this.cache.delete(request.url);
-      const resource = this.resources.get(request.url);
-      if (resource) {
-        resource.emit('delete');
-      }
-    }
-
-    return response;
-
-  };
 
 }
